Fix overflow in the filter that occured when I used an amplifier for the antenna by scaling `w' down by 5 bits, and using 64-bit arithmetic for the a1 and a2 coefficients. Now we have quite some headroom for the amplitude to become bigger, and if we observe that overflows are happening again, we can always increase the number of bits with which `w' is scaled down.
Because `w' is smaller, the output of the filter is now also always so small that we can
store it in a 16-bit buffer instead of a 32-bit buffer.

Increased the Q factor of the filter to 15000 while we're at it, since then we have
a nice balance between the quantization error of coefficients a1 and a2 (they are both
relatively small).