Change capture depth to 20000 samples, so that the sample rate becomes exactly 100Hz for the averages. This makes sure that the the 1s period fits in a whole number of averages (we could have also done this by changing the upsampling factor we difined as 4 before to 512, but then we would have a way higher sample rate for the averages than desirable (15.625kHz).
The reason that I'm doing this is so that we can later possibly use a matched filter
where we can use convolution to detect the exact phase over the course of several
seconds/minutes, and the convolution kernel will align exactly to one period of
the DCF77 signal (1s long).

This change of the capture depth to a number that is not a power of 2 does mean that
the divisions by the capture depth don't simply optimize to bit shifts, in one place
this means that it is slightly faster to use the asynchronous feature of the hardware
divider in the RP2040.