Add back a very simple pulse length detection algorithm, so that we can process the bits again. I already see massive improvements regarding the bit level detection compared to when we were just measuring the pulse length with some threshold.
The sync pulse detection isn't very robust yet, so it might be good to do some phase locking
to that too, since we know the period is exactly 60s always (except when a leap second is introduced).

Also heavily optimized the wrapping and summing of the convolution output buffer (not
using modulo (which takes 8 cycles) each iteration).