Use the cool GNU C feature of statements and declarations in expressions for the `MAX_IDX' macro, since that makes its usage more function-like (so that it is not unclear anymore that it is assigning a value to the `max_idx' argument).
Also make sure that we don't change phase (`max_idx') for the start of a window detection very
close to the start of a window, since then it would be possible that the `start_100ms',
`start_200ms' and `end_200ms' booleans are true for more than one sample in a second, or
not one sample at all in a second. Do this by only updating the used phase (`prev_max_idx') when
we are in the middle of the 800ms period (when `conv_buf_idx_mod_1s' is about the same as `max_idx').

Furthermore pass the timestamp of the previous start of a 100ms window to the `process_bit' function
because we use averages from 1 and 2 windows ago there (essentially the bit 1s ago,
so timestamp has to be 1s ago too).