Count the number of times the leap announcement bit is set in an hour, if this is larger than some threshold, we set the number of seconds in a minute to 61 (instead of 60) at the whole last minute of the hour. Now we only need to determine with some degree of robustness when the last minute of the hour is. We can probably do this more easily later when we have synchronized the RTC.
Also there is still an error in the way the phase of the sync mark was applied to the bit number,
since it won't work when `max_idx' is 59.

All of the above was done by writing a little bit of test code around the `process_bit' function
with data for when a leap second occurred taken from https://dcf77logs.de/.

Add previously missing break statement for bit 16 (DST switch announcement bit) as well as
explicit fallthrough annotation with the double square bracket attribute notation instead of
writing it in a comment to prevent warnings (which is kind of insane; just the fact that GCC parses
comments to determine its behaviour). For this we do need the to set the C standard version to
GNU23.