Wrap the large buffer in which the averages are stored instead of wrapping the convolution output buffer. This is a lot faster, because the convolution doesn't have to be applied to a large buffer now, and it is mathematically the same since convolution is LTI.
`avg_buf' can also be made 16-bit instead of 32-bit because the averages never go near 2^15-1.
This in turn makes it possible to double the size of `avg_buf' (so store 5 minutes instead of 2.5 minutes).

It now takes about 2.2ms to convolve the wrapped buffer (with 1s worth of data).