Keep track of parity errors for all fields so we can do something with them later (instead of only printing them). Also apply the same technique as I used for the leap second announcement bit to the DST switch announcement bit (check how often it is 1 during the hour, if it's more than 30, we can be pretty sure that we will switch).
Also fix an issue with how the last second of the hour was detected (before we had to set `last_minute'
to true at the last second of the second to last minute as well, but this is not required anymore.
Do this by wrapping `bit' around to 0 at the beginning of the `process_bit' function, instead of at the end.
There was also some undefined behaviour in accessing the `bit_after_sync' variable (since we were accessing
it while it was uninitialized, so just getting the value of the previous call to `process_bit',
this is why we had to check if `bit_after_sync == seconds_in_minute - 2' instead of `bit_after_sync == seconds_in_minute - 1'.

Test all of the above in the `leap_second_test.c' script as well.