Fix two bugs, the first is that the `synced_timestamp_offset' was added to `prev_timestamp' instead of `timestamp', this caused 1s long 10ms peaks in the timestamp used in `process_bit' (on top/instead of it gradually drifting). The reason that `synced_timestamp_offset' has to be added to `timestamp' instead of `prev_timestamp' is that when `synced_max_idx' changes, `start_100ms' also changes, and there is a value for `timestamp' associated to this, so we have to access the `timestamp' at the same time as `synced_timestamp_offset' and store it in `prev_timestamp' instead of adding `synced_timestamp_offset' 1s earlier, when `prev_timestamp' hasn't changed to the timestamp associated with the new `start_100ms' yet.
The second is that it can take a very long time for the RTC to be in sync if the first
received date/time causes an invalid XOR mask for day or month. This is because `last_rtc_alarm_datetime'
gets initialized to all zeroes (which is in itself already an invalid date since day and month cannot be 0),
so the `rtc_set_datetime' function doesn't set any of the fields, which breaks the algorithm,
so then it is just a case of luck until both day and month are flipped by the XOR mask at the same time.
Fix this by setting the day and month to 1 in an initializer, also set the year to 2000 while we're at it so
that we don't get negative numbers when subtracting 2000 in `compare_time'. (This resulted in the year
being 2080 after converting to and from BCD.)